Event service split into abstraction and implementation
